Add highscore text if the game was won
This commit adds a highscore text if the game was won. If the game was
won for the first time or a new highscore is reached, it will display a
text congratulating the player. Otherwise it wil display the users
personal best. The highscores are saved separately for every game
mode (combination of number of colors and board size).